# package/openapi.yaml
openapi: 3.1.0
info:
  title: My Tasco Secure Agentic RAG API
  version: 1.0.0
  description: |
    Full public contract for the current secure RAG prototype and planned Agentic RAG platform.
    Every operation declares x-implementation-status. Planned operations are contracts, not claims
    that the current backend implements them. OpenAI is an internal outbound dependency and is
    never proxied directly to clients.
  contact:
    name: My Tasco AI Workspace Team
servers:
  - url: http://127.0.0.1:8000
    description: Local prototype
  - url: https://hackathon.example.com/mytasco/v1
    description: Configurable deployment facade
...
paths:
...
  /mytasco/v1/aiwsp/users:
    get:
      tags: [Legacy]
      operationId: listDemoUsers
      summary: List synthetic users
      description: Development/demo inventory only; this endpoint must be disabled in production.
      x-implementation-status: implemented-demo-only
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200': {description: Synthetic user inventory, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /mytasco/v1/aiwsp/knowledge/search:
    post:
      tags: [Legacy]
      operationId: searchKnowledge
      summary: ACL-aware enterprise knowledge search
      x-implementation-status: implemented-core
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/AppCode'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SearchRequest'}
      responses:
        '200': {description: Authorized search results, content: {application/json: {schema: {$ref: '#/components/schemas/SearchEnvelope'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /mytasco/v1/aiwsp/assistant/chat:
    post:
      tags: [Legacy]
      operationId: legacyChat
      summary: Legacy synchronous secure RAG chat
      description: Compatibility endpoint. New clients should use chat sessions and runs.
      x-implementation-status: implemented-core
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/AppCode'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LegacyChatRequest'}
      responses:
        '200': {description: Validated grounded answer, content: {application/json: {schema: {$ref: '#/components/schemas/LegacyChatEnvelope'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /mytasco/v1/aiwsp/index/rebuild:
    post:
      tags: [Legacy, Documents]
      operationId: rebuildIndex
      summary: Rebuild the prototype vector index
      x-implementation-status: implemented-demo-only
      responses:
        '200': {description: Rebuilt index information, content: {application/json: {schema: {$ref: '#/components/schemas/GenericEnvelope'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    AppCode: {type: apiKey, in: header, name: X-App-Code, description: Must equal MYTASCO}
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: {type: string, format: uuid}
    AppCode:
      name: X-App-Code
      in: header
      required: true
      schema: {type: string, const: MYTASCO}
...
  responses:
    InvalidRequest: {description: Invalid request, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    Unauthorized: {description: Missing or invalid authentication, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    Forbidden: {description: Authenticated but not authorized, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
    RateLimited: {description: Rate or budget limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    ServiceUnavailable: {description: Required secure dependency unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  schemas:
    Classification:
      type: string
      enum: [Public, Internal, Confidential, Restricted]
...
    PageInfo:
      type: object
      required: [pageSize, currentPage, totalRecord]
      properties:
        pageSize: {type: integer, minimum: 1}
        currentPage: {type: integer, minimum: 0}
        totalRecord: {type: integer, minimum: 0}
    ErrorEnvelope:
      type: object
      required: [status, code, message, requestId]
      properties:
        status: {type: string, const: error}
        code: {type: string, enum: [invalid_request, unauthorized, forbidden, not_found, conflict, expired, timeout, rate_limited, service_unavailable, internal_error]}
        message: {type: string}
        requestId: {type: string}
    GenericEnvelope:
      type: object
      required: [status, message, body, requestId]
      properties:
        status: {type: string, const: success}
        message: {type: string, const: SUCCESS}
        body: {}
        requestId: {type: string}
...
    SearchRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 8000}
        filters:
          type: object
          additionalProperties: false
          properties:
            department: {type: string}
            classification: {$ref: '#/components/schemas/Classification'}
        pageInfo:
          type: object
          properties:
            pageSize: {type: integer, minimum: 1, maximum: 20, default: 10}
            currentPage: {type: integer, minimum: 0, default: 0}
    SearchHit:
      type: object
      required: [documentId, chunkId, title, department, classification, section, snippet, score]
      properties:
        documentId: {type: string}
        chunkId: {type: string}
        title: {type: string}
        department: {type: string}
        classification: {$ref: '#/components/schemas/Classification'}
        section: {type: string}
        snippet: {type: string}
        score: {type: number, minimum: 0, maximum: 1}
        allowedReason: {type: string}
        piiTypes: {type: array, items: {type: string}}
    SearchEnvelope:
      allOf:
        - {$ref: '#/components/schemas/GenericEnvelope'}
        - type: object
          properties:
            body:
              type: object
              properties:
                result: {type: array, items: {$ref: '#/components/schemas/SearchHit'}}
                pageInfo: {$ref: '#/components/schemas/PageInfo'}
                permissionDecision: {type: string}
                likelyPermissionDenied: {type: boolean}
                deniedMatchCount: {type: integer}
                dlp: {type: object, additionalProperties: true}
                roleIdentifier: {type: object, additionalProperties: true}
    LegacyChatRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message: {type: string, minLength: 1, maxLength: 8000}
        conversationId: {type: [string, 'null']}
        mode: {type: string, default: knowledge}
        stream: {type: boolean, default: false}
    Citation:
      type: object
      required: [documentId, title, section, classification]
      properties:
        evidenceId: {type: string}
        documentId: {type: string}
        version: {type: string}
        title: {type: string}
        section: {type: string}
        page: {type: [integer, 'null']}
        classification: {$ref: '#/components/schemas/Classification'}
    LegacyChatEnvelope:
      allOf:
        - {$ref: '#/components/schemas/GenericEnvelope'}
        - type: object
          properties:
            body:
              type: object
              properties:
                conversationId: {type: string}
                answer: {type: string}
                citations: {type: array, items: {$ref: '#/components/schemas/Citation'}}
                permissionDecision: {type: string}
                confidence: {type: string}
                modelUsed: {type: string}
                redactionApplied: {type: boolean}
                dlpCategories: {type: array, items: {type: string}}
                retrieval: {type: object, additionalProperties: true}
                roleIdentifier: {type: object, additionalProperties: true}
...
security:
  - BearerAuth: []
    AppCode: []